requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
requests==2.31.0
gunicorn==20.1.0
bing-image-downloader==1.0.4
APScheduler==3.10.4
python-dotenv==1.0.0